fix controller

+++ src/main/java/org/example/tddrps_game/repository/GameRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GameRepo extends CrudRepository<Game,Long> {
+public interface GameRepo extends CrudRepository<Game,Integer> {
 }
